Organize mobile screens and update paths
