+++ mobile/App.tsx
@@ -3,7 +3,7 @@ import "react-native-gesture-handler";
 import { NavigationContainer } from "@react-navigation/native";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import { NativeBaseProvider, extendTheme } from "native-base";
-import RecordScreen from "./screens/RecordScreen";
+import RecordScreen from "./screens/home/RecordScreen";
 
 // ──────────────────────────────────────────────────────────────────────────────
 // Theme (minimal)
